+++ src/get.rs
@@ -248,6 +248,10 @@ impl Fetcher {
     ) -> bool {
         use minisign_verify::{PublicKey, Signature};
 
+        if config.sigs.is_empty() {
+            panic!("{} for index '{}'. Please add one with --pub-key or use --no-verify.", err_color_print("No public key(s)", PossibleColor::BrightRed), config.index);
+        }
+
         let sig = &self.fetch_str(id, version, sig_file);
         let signature = Signature::decode(sig).expect("Signature was malformed.");