+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## Upcoming
 
+## [0.6.4](https://github.com/cargo-prebuilt/cargo-prebuilt/releases/tag/v0.6.4)
+
 - Internally removed a lot of panics for better reliability.
 - CLI color change.
 - Panic on empty public keys when --no-verify is not used.
-- Added GitHub Private index. (beta)
-
-## [0.6.4](https://github.com/cargo-prebuilt/cargo-prebuilt/releases/tag/v0.6.4)
+- Added GitHub Private index support. (beta)
 
 ## [0.6.3](https://github.com/cargo-prebuilt/cargo-prebuilt/releases/tag/v0.6.3)